Fix issue in Rate Limit where delete button was disabled when it was supposed to be enabled after deleting the last Security Policy connection
